Re-factored Nav Bar to re-usable Angular component
