+++ app/src/components/components.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { NavbarComponent } from './navbar/navbar';
+@NgModule({
+	declarations: [NavbarComponent],
+	imports: [],
+	exports: [NavbarComponent]
+})
+export class ComponentsModule {}
