+++ app/src/components/navbar/navbar.scss
@@ -1,0 +1,3 @@
+navbar {
+
+}
