+++ app/src/pages/account-change-password/account-change-password.html
@@ -1,11 +1,5 @@
 <ion-header>
-  <ion-navbar>
-    <!--<button menuToggle>
-      <ion-icon name="menu"></ion-icon>
-    </button>
-    -->
-    <ion-title>Change Password</ion-title>
-  </ion-navbar>
+  <navbar title="Change Password"></navbar>
 </ion-header>
 
 <ion-content padding>
