+++ app/src/pages/account-confirmation-code/account-confirmation-code.html
@@ -1,9 +1,6 @@
 <ion-header>
-  <ion-navbar>
-    <ion-title>Confirm Sign-Up</ion-title>
-  </ion-navbar>
+  <navbar title="Confirm Sign Up"></navbar>
 </ion-header>
-
 
 <ion-content padding>
