+++ app/src/pages/account-forgot-password/account-forgot-password.html
@@ -1,11 +1,5 @@
 <ion-header>
-  <ion-navbar>
-    <!--<button menuToggle>
-      <ion-icon name="menu"></ion-icon>
-    </button>
-    -->
-    <ion-title>Password Reset</ion-title>
-  </ion-navbar>
+  <navbar title="Reset Password"></navbar>
 </ion-header>
 
 <ion-content padding>
