+++ app/src/pages/account-signin/account-signin.html
@@ -1,17 +1,6 @@
 <ion-header>
-  <ion-navbar>
-    <!--<button menuToggle>
-      <ion-icon name="menu"></ion-icon>
-    </button>-->
-    <ion-buttons right>
-      <button ion-button icon-left [navPush]="accountSignupPage">
-        <ion-icon name="contact"></ion-icon>Register
-      </button>
-    </ion-buttons>
-    <ion-title>Sign In</ion-title>
-  </ion-navbar>
+  <navbar title="Sign In"></navbar>
 </ion-header>
-
 
 <ion-content padding>
