+++ app/src/pages/account-signup/account-signup.html
@@ -1,12 +1,6 @@
 <ion-header>
-  <ion-navbar>
-    <!--<button menuToggle>
-      <ion-icon name="menu"></ion-icon>
-    </button>-->
-    <ion-title>Register</ion-title>
-  </ion-navbar>
+  <navbar title="Register"></navbar>
 </ion-header>
-
 
 <ion-content padding>
