+++ app/src/pages/location-add/location-add.html
@@ -1,12 +1,6 @@
 <ion-header>
-  <ion-navbar>
-    <!--<button menuToggle>
-      <ion-icon name="menu"></ion-icon>
-    </button>-->
-    <ion-title>Add a Location</ion-title>
-  </ion-navbar>
+  <navbar title="Add a Location"></navbar>
 </ion-header>
-
 
 <ion-content padding>
 
@@ -29,7 +23,7 @@
       </ion-item>
       <p [hidden]="description.valid || !submitted" color="danger" padding-left>
         Description is required
-      </p>      
+      </p>
 
       <ion-item>
         <ion-label floating>Image URL</ion-label>
@@ -37,7 +31,7 @@
       </ion-item>
       <p [hidden]="imageUrl.valid || !submitted" color="danger" padding-left>
         Image URL is required
-      </p>     
+      </p>
 
     </form>
   </ion-list>
